Potential update to travis-ci yaml configuration so that both teensy and pro micro variants are built
For Teensy, due to the limitations of how the build system integration has to be configured, in order to apply custom pre-processor macros at build time, we need to hard code the compiler arguments in boards.txt, which means we effectively map the teensy 4.0 to a specific device type.

We could workaround this by effectively creating a new fake board that is a copy of teensy40, but it's not the best solution.

The Pro micro uses arduino-cli, which in theory, should "just work"