Button bindings
Buttons now have bindings, these bindings are loaded and saved in the eeprom.
On startup, the micro controller will check to see if there are valid bindings stored on the eeprom by checking a CRC.
If there isn't a valid CRC, then a default configuration is saved onto the eeprom.

Also added a debug print library to make printing to the console easier than using Serial.print