Some refactoring to enable control of the LEDs
There is now an LED "matrix" class, which is meant to be a companion to the Button "matrix". There is also an LED controller, which will allow for LEDs not controlled by a MAX chip/matrix.

A similar refactor will need to happen to buttons so we have a "button controller" in addition to the matrix